+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- (NOTIFIER-39) Fixed Cart Simulation logic when verifying that a sku can ship to a shopper
+
 ## [1.7.3] - 2022-05-16
 
 (NOTIFIER-31) Changed forwarding address